requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
